removed saving and checking accounts classes, updated dbcontext
