--- eMNY/eMNY.Domain/Interfaces/IAccount.cs
@@ -1,8 +1,0 @@
-namespace eMNY.Domain.Interfaces
-{
-    public interface IAccount
-    {
-      int Id { get; }
-      bool IsValid();
-    }
-}
