+++ eMNY/eMNY.Domain/eMNY.Domain.csproj
@@ -7,8 +7,4 @@
   <ItemGroup>
     <Folder Include="Abstracts\" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Compile Remove="Models\Account.cs" />
-  </ItemGroup>
 </Project>
